create_guestimg.sh: Adjust the network settings for the pkvm project
This patch changes tghe network to 192.168.10.x. I already published a
script on the wiki page that used the 192.168.10.x network, so let's
use it here too.

I had some problems that I believe may have come from network manager, so
let's remove it. It doesn't make sense to have network manager installed
if we are using /etc/network/interfaces.

I have an impression that the removal of the network manager may cause
the /etc/resolv.conf file to be updated to some symbolic link to some
systemd stuff, so therefore let's recreate it at the end.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>