scripts/create_guestimg.sh: Increase default systemd timeout
Because the guest will run so slowly with emulation and nested
virtualization, we will increase this timeout, so that systemd will wait
longer for devices. This is especially important for ttyS0, which is used
as a console and its timeout would cause getty to fail.

Unfortunately, or perhaps fortunately, this knob also changes the default
timeout not only for devices but for everything else as well. This could
be both good and bad.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>