scripts/create_guestimg.sh: Increase timeouts for ifupdown-pre.service
Again, because of emulation and nested virtualization, we need to increase
the timeout of ifupdown-pre.service. Here we need to not only increase
the TimeoutStartSec but also increase the timeouts of the command that
is called by this job, i.e. the "udevadm settle", needs to be called
with a higher timeout, otherwise that command will run out of time.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>