Add targets for creating additional guestimages
We need to test pKVM with more than one guest image. For this reason three
new targets are added: guestimage2 guestimage3, guestimage4. These are not
built with the all target. These are just modified versions of the regular
guestimage with a different hostname and a different IP address.

This target will first duplicate the guestimage and then modify it so that
it has a different hostname and uses a different network address.

Also modify the run-crosvm.sh script so that, it's possible to run more
than one guest simultaneously. The script need to setup a different disk
image, subnet, tap device, and socket for each guest.

There are three new arguments:
-i NR	  	 This is a convenience. By giving a number between 0..3.
   		 You can modify the default values for disk image, subnet,
		 tap device, and socket. 0 is the value for the first
		 guest, 1 is the second, 2 is the third, and 3 for the
		 fourth. The network will be between, 192.168.10 and
		 192.168.13. The default disk image will be
		 ubuntuguest.${network}.qcow2

-n subnet	 This specifies the subnet to be used.

-s socket	 This specifies the socket to be used.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>