Build the Ubuntu system only once instead of trice
The target-crosvm, hostimage,  guestimage all build an ARM64 Ubuntu system
almost from scratch so that there is a lot of repeated work when building
the 'all' target. The worst part is that all of them do the
'apt-get dist-upgrade' command which takes a lot of time.

Thus, add a new ubuntu-template target that builds and Ubuntu template
that can then be reused by the target-crosvm, guestimage, and hostimage
targets.

In order to rebuild the ubuntu-template, it's necessary to do a
an 'make ubuntu-template-distclean', or possibly 'make distclean' to do
a more thorough cleaning of more all targets.

The purpose of this change is only to improve the build time, not to
provide any new feature, except that a new distclean target that
will do a more thorough clean.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>