Make the clean and distclean targets more consistent
The idea behind this chane is that we shave of the build time with the
following sequence

make all
make clean
make all

With this change, the last 'make all' will not rebuild the tools and
ubuntu-template target.

If a user want to force  the rebuilding of the ubuntu-template and the
tools, then the 'make distclean' should be used instead of 'make clean'.

The 'make tools-distclean' and 'make ubuntu-template-distclean' can be
used for more fine grained control, to only force the rebulding of one
of the targets.

This change also make the 'make tools' target idempotent, in the sense
that it will not try to rebuild anything if it detects the presence
of a cross gcc and gdb in the tools directory.

The reason for all this is to make it easier to reduce build times.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>