Userspace hypdbg utility on rust
The utility rewritten on rust for the purposes of readability
and manageability. There are default values of options and an argument
as well as short descriptio about their usage.

The resulting executable footprint take ~14Mb of memory which, author
assumes, is still acceptable for the emulated and hw pkvm setups

Signed-off-by: Grigoriy Romanov <grigoriy.romanov@unikie.com>