guest-kernel: Update to 6.6.40
This updates the guest kernel to 6.6.40 and also fixes a minor bug in
scripts/guest-kernel-patch-fiddle.sh; it should always use PATCH_DIR
variable, not hardcoded.

Because of the kernel upgrade, this seems to fix a bug where networking
stopped working in the guest after some tens or hundreds of megabytes.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>