guest-kernel: Improve the kernel configuration of the guest kernel
The guest kernel build currently builds a lot of kernel modules that are
not needed, e.g. support for HW that we don't have in the guest, etc.

For this reason, let's create a new guest_defconfig, which is adapted
for our guest build. The first step is to disable modules that are
obviously not needed. It is likely that even more features could be
disabled in the kernel but that would require more work.

The big benefit with not building these modules is that it speeds up the
kernel build considerably.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>