linux-host: Update to get the reduced qemu_defconfig
We update the host kernel to get these new commits:
5a1dbed63b7a qemu_defconfig: Disable kernel modules that are not needed
74b197732749 Add debug_defconfig

The debug_defconfig is currently not in use, so that commit doesn't
matter currently. What we want is the updated qemu_defconfig.

An obvious piece of work that remains is to refresh the debug_defconfig so
that it is based on the reduced qemu_defconfig, instead of the old one.
Right now, if you use debug_defconfig, you will build a lot of
unnecessary modules.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>