hack: Fix build for Ubuntu 24.04
This is hack to fix the build for Ubuntu 24.04. With this version, we hit
the snag that the ELF loader of the new qemu-user-static is not compatible
with loading old static PIE images of the Ubuntu 22.04 images that we want
to build. This is because there is apparently some kind of conflict in the
initialization of the GLIBC library with the way the new qemu handles the
initializtion of the program break (brk) when loading an ELF image. The
symptom is that ldconfig, which is a static PIE binary segfaults when being
called from a post installation script by dpkg when trying to configure
the libc package.

It would likely be possible to patch the qemu to handle both old and
new static PIE binaries correctly but this was deemed too much work at
this point.

Therfore, what we are doing here is to:
1. introuce a new build target qemu-user, which builds an older version of
  qemu. We repurpose the disused qemu source code tree for this purpose.
  Now te ./oss/qemu source tree is used to build the system level qemu that
  is used when running our images and the ./qemu source tree is used to
  build this qemu-user-static.

2. In order to get the correct old version of qemu-user-static, which is
   v7.2.14, we move the ./qemu source tree to this version.

3. The following build targets are modified to utilize this custom
   qemu-user-static by means of registering it with the binfmt_misc kernel
   module: ubuntu-template, target-crosvm, hostimage, guestimage. This is
   necessary because when dpkg is trying to run a post installation script,
   it will fork and exec and the interpreter used will then be the one
   that is registered with the binfmt_misc service.

This hack can possibly be removed if we upgrade our host and guest images
to be based on something newer. On the other hand we may then want to
modify it to instead build a newer qemu-user-static for those who are using
a distro with an older qemu-user-static. The compatibility of an old
qemu-user-static with never glibc has not been investigated.

If you are still using Ubuntu 22.04 or any distro with an older qemu, older
than v8.2, then you may want to forego building qemu-user-static and you
can do so by using the variable BUILD_QEMU_USER, e.g.:

make BUILD_QEMU_USER=0 all 2>&1 |tee buildlog-all.txt

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>