linux-host: Update to a 6.6.73 based pKVM kernel
We upgrade the host kernel to a kernel that is based on mainline 6.6.73.
This kernel has been patched with the pKVM patches from
android15-6.6.66_r00.

A new branch legacy-linux-5.15 has been created. We will probably not
maintain that but it's added as a convenience to be able to quickly
checkout the old 5.15.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>