linux-host: Update to a new branch of the 6.6.73 based pKVM kernel
The old branch contained a bad fix for a critical issue on the Nvidia
Orin AGX right at the beginning, i.e. after the last upstream commit. We
could revert the bad fix and add a proper fix but that would make the old
branch unuitable for using git bisect, which can be a critical
capability.

For this reason, a new branch have been created for Tegra. Basically,
the difference is that this commit:
a0070a011c07 ("HACK: cpu_pm: work around non-working cpudile for EL1")

Has been replaced by this:
949c70ca09a3 ("cpuidle: psci: disable non-WFI states for EL1 on Tegra")

This doesn't really do anything for running pKVM on QEMU but we want
to be in sync with what is done on the Nvidia AGX Orin.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>