linux-host: Again update to a new branch of the 6.6.73 base pKVM kernel
The previous branch tried to fix the idle issue on the Tegra with this:
949c70ca09a3 ("cpuidle: psci: disable non-WFI states for EL1 on Tegra"

This bad fix was applied at the beginning of the old branch, and this might
make git bisect usage more difficult.

In the new branch we have removed this fix, not replaced it, and the new
fix sits in the ARM Trusted Firmware.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>