scripts/crate_*img.sh: Fix race conditions when connecting images to nbd
These scripts are two types of race conditions. There isn't really any
guarantee that a device file like /dev/nbd0, or /dev/nbd0p1 has appeared
because their appearance is not really synchronous with respect to the
sheell script.

This above type of race is solved by using "udevadm wait" or if it is not
available by using udevadm settle in a poll loop.

The second issue is that "qemu_nbd --connect" and "qemu-nbd --disconnect"
may return before the file in question has been connected or diconnected.

This race is solved by polling the size of the /dev/nbd0. If it is 0, then
we assume that the file is disconnected from the device, if it is greater
than 0, then we assume that it is connected.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>