linux-host: Update to the latest commit in branch linux-6.6.y-pkvm5
There are two new commits in the branch:
c65ac8ed295e HACK: Workaround to get "protected mode with firmware" to work
b389232f4934 PKVM: add support for guest to guest memory share

I really wish that we would not need
c65ac8ed295e ("HACK: Workaround to get "protected mode with firmware" to work")
but after testing for the first time with the kernel integrity checker, it
seems like it is necessary to be able to boot the guest in protected mode
with firmware.

I always tested protected without firmware before and in
that case the MMIO Guard is apparently not enabled, so the problem is not
seen.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@tii.ae>